# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "financialPortfolioApiBackend",
        "description": "API Documentation for your application.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development Server"
        },
        {
            "url": "http://localhost",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/transactions/transfer": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Perform a money transfer between users",
                "operationId": "3c8a4f994fb8ebaae0ed746708e69327",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransferRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transfer performed successfully or denied with a reason.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Transferência realizada com sucesso!"
                                        },
                                        "transaction": {
                                            "$ref": "#/components/schemas/TransactionResponse"
                                        },
                                        "reason": {
                                            "description": "Reason for denial if status is DENIED.",
                                            "type": "string",
                                            "example": "Insufficient balance",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - missing or invalid token.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - user type not allowed to transfer.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuários do tipo Lojista não podem realizar transferências."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Tipo de usuário não permitido."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payee not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Recebedor não encontrado."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "O ID do usuário recebedor não corresponde a um usuário existente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error - invalid input data.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao realizar a transferência."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Ocorreu um erro inesperado ao processar a transferência. Por favor, tente novamente mais tarde."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/transactions/deposit": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Perform a money deposit for the authenticated user",
                "operationId": "50765c5eda03030a06d3a6d9635e08e4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DepositRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Deposit performed successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Depósito realizado com sucesso!"
                                        },
                                        "transaction": {
                                            "$ref": "#/components/schemas/TransactionResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - missing or invalid token.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error - invalid input data.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao realizar o depósito."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Ocorreu um erro inesperado ao processar o depósito. Por favor, tente novamente mais tarde."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DepositRequest": {
                "title": "Deposit Request",
                "description": "Data required for a user to make a deposit.",
                "required": [
                    "amount"
                ],
                "properties": {
                    "amount": {
                        "description": "The amount to be deposited. Must be greater than zero.",
                        "type": "number",
                        "format": "float",
                        "example": 100.5
                    }
                },
                "type": "object"
            },
...
            "TransferRequest": {
                "title": "Transfer Request",
                "description": "Data required for a user to initiate a money transfer.",
                "required": [
                    "payee_id",
                    "amount"
                ],
                "properties": {
                    "payee_id": {
                        "description": "The UUID of the user who will receive the transfer. Must be an existing user and cannot be the payer's own ID.",
                        "type": "string",
                        "format": "uuid",
                        "example": "987e6543-d2c1-b0a9-8765-43210fedcba9"
                    },
                    "amount": {
                        "description": "The amount to be transferred. Must be a positive number greater than zero.",
                        "type": "number",
                        "format": "float",
                        "example": 250.75
                    }
                },
                "type": "object"
            },
...
            "AuthenticationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Falha na autenticação."
                    },
                    "error": {
                        "type": "string",
                        "example": "Credenciais inválidas."
                    }
                },
                "type": "object"
            },
            "TransactionResponse": {
                "description": "Schema representing a transaction, including details about sender and recipient.",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "d3f4e5c6-7890-1234-5678-abcdefabcdef"
                    },
                    "payer_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "a1b2c3d4-5678-90ef-1234-567890abcdef"
                    },
                    "payee_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "b2c3d4e5-6789-01ab-2345-678901abcdef"
                    },
                    "amount": {
                        "type": "string",
                        "format": "decimal",
                        "example": "150.00"
                    },
                    "type": {
                        "type": "string",
                        "example": "TRANSFER"
                    },
                    "status": {
                        "type": "string",
                        "example": "REVERSED"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-01T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-02T15:00:00Z"
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Erro de validação."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}